Fixes for testsuite, new tests (pin all, unpin)
Fix bug about using relative paths for index configuration in the python setup.
