+++ testsuite/tests/pin/all/my_index/index/he/hello2.toml
@@ -1,0 +1,8 @@
+[general]
+description = """"Hello, world!" demonstration project"""
+licenses = []
+maintainers = ["example@example.com"]
+maintainers-logins = ["mylogin"]
+
+['0.1']
+origin = "file://."
