+++ testsuite/tests/pin/all/my_index/index/index.toml
@@ -1,0 +1,1 @@
+version = "0.2"
