+++ testsuite/tests/pin/all/test.yaml
@@ -1,0 +1,4 @@
+driver: python-script
+indexes:
+    my_index:
+        in_fixtures: false
