+++ testsuite/tests/workflows/init-with-pin/test.py
@@ -6,20 +6,7 @@ import os
 
 from drivers.alr import run_alr
 from drivers.asserts import assert_eq
-
-
-def check_line_in(filename, line):
-    """
-    Assert that the `filename` tetx file contains at least one line that
-    contains `line`.
-    """
-    with open(filename, 'r') as f:
-        for l in f:
-            if l.rstrip() == line:
-                break
-        else:
-            assert False, 'Could not find {} in {}'.format(
-                repr(line), filename)
+from drivers.helpers import check_line_in
 
 
 # Create a new "xxx" program project
@@ -40,9 +27,10 @@ with open('xxx.gpr', 'w') as f:
     f.write('with "libhello";\n')
     f.write(content)
 
-# Pin the version of libhello
-run_alr('pin')
-check_line_in(session_file, 'libhello = "=1.0.0"')
+# Pin the version of libhello and verify pin is there
+run_alr('pin', 'libhello')
+p = run_alr('pin')
+assert_eq('libhello 1.0.0\n', p.out)
 
 # Build and run "xxx"
 with open(os.path.join('src', 'xxx.adb'), 'w') as f:
